0909/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.7
django-appconf==1.0.4
django-bootstrap-v5==1.0.5
django-imagekit==4.0.2
pilkit==2.0
Pillow==8.3.2
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
